requirements.txt
# Core dependencies
streamlit==1.32.0
Pillow==10.2.0
svgwrite==1.4.3
networkx==3.2.1
nltk==3.8.1
scikit-learn==1.4.0
numpy==1.26.4
matplotlib==3.8.2

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Audio processing (for podcast generation)
gTTS==2.5.1
pyttsx3==2.90
soundfile==0.12.1

# Video generation
moviepy==1.0.3
opencv-python==4.9.0.80

# Presentation generation
python-pptx==0.6.23

# Image processing
Pillow==10.2.0
cairosvg==2.7.1

# UI enhancements
streamlit-option-menu==0.3.12
streamlit-custom-notification-box==0.1.1

# Data processing
pandas==2.2.0
scipy==1.12.0

# Text processing
transformers==4.37.2
torch==2.2.0
spacy==3.7.4

# Visualization
plotly==5.18.0
seaborn==0.13.2

# Development tools
python-dotenv==1.0.1
tqdm==4.66.2